master

+++ helper/src/main/java/com/vsg/helper/common/util/viewModel/IViewModelAllSimpleListIdRelation.kt
@@ -3,5 +3,5 @@ package com.vsg.helper.common.util.viewModel
 import com.vsg.helper.common.model.IEntity
 
 interface IViewModelAllSimpleListIdRelation<TEntity> where TEntity: IEntity {
-    fun viewModelViewAllSimpleList(idRelation: Long): List<TEntity>?
+    fun viewModelViewAllSimpleList(idRelation: Int): List<TEntity>?
 }
