+++ helper/src/main/java/com/vsg/helper/common/util/viewModel/IViewModelHasItemsRelationType.kt
@@ -1,5 +1,5 @@
 package com.vsg.helper.common.util.viewModel
 
 interface IViewModelHasItemsRelationType<TFilter> where TFilter : Enum<TFilter> {
-    fun viewModelViewHasItems(idRelation: Long, filter: TFilter): Boolean
+    fun viewModelViewHasItems(idRelation: Int, filter: TFilter): Boolean
 }
